requirements.txt
deepface==0.0.75
mediapipe
fastapi==0.92.0
numpy
opencv_contrib_python==4.7.0.68
Pillow==9.4.0
uvicorn==0.20.0
python-multipartgit